refactor: fix calendar to compound-component
